se_results_count / items_count の型を i32 から i64 に変更する

+++ src/entity/backlinks/anchor.rs
@@ -5,6 +5,6 @@ use serde::{Deserialize, Serialize};
 pub struct BacklinksApiAnchor {
     pub target: Option<String>,
     pub total_count: Option<i32>,
-    pub items_count: Option<i32>,
+    pub items_count: Option<i64>,
     pub items: Option<Vec<BacklinksElementBacklinkAnchor>>,
 }
