+++ src/entity/serp/google/news.rs
@@ -20,8 +20,8 @@ pub struct SerpApiGoogleNewsTask<T> {
     pub spell: Option<SerpApiGoogleOrganicTaskSpell>,
     pub refinement_chips: Option<SerpApiElementRefinementChips>,
     pub item_types: Option<Vec<String>>,
-    pub se_results_count: Option<i32>,
-    pub items_count: Option<i32>,
+    pub se_results_count: Option<i64>,
+    pub items_count: Option<i64>,
     pub items: Option<Vec<T>>,
 }
